+++ Samples/iOS-Swift/iOS-Swift/ErrorsViewController.swift
@@ -13,13 +13,20 @@ class ErrorsViewController: UIViewController {
     override func viewDidLoad() {
         super.viewDidLoad()
         addDSNDisplay(self, vcview: dsnView)
+
+        if SentrySDKOverrides.Feedback.useCustomFeedbackButton.boolValue {
+            let button = SentrySDKWrapper.shared.feedbackButton
+            view.addSubview(button)
+            button.leadingAnchor.constraint(equalTo: view.safeAreaLayoutGuide.leadingAnchor, constant: 8).isActive = true
+            button.bottomAnchor.constraint(equalTo: view.safeAreaLayoutGuide.bottomAnchor, constant: -8).isActive = true
+        }
     }
 
     override func viewDidAppear(_ animated: Bool) {
         super.viewDidAppear(animated)
         SentrySDK.reportFullyDisplayed()
         
-        if ProcessInfo.processInfo.arguments.contains("--io.sentry.feedback.inject-screenshot") {
+        if SentrySDKOverrides.Feedback.injectScreenshot.boolValue {
             NotificationCenter.default.post(name: UIApplication.userDidTakeScreenshotNotification, object: nil)
         }
     }
@@ -84,6 +91,16 @@ class ErrorsViewController: UIViewController {
         fatalError("This is a fatal error. Oh no 😬.")
     }
 
+    @IBAction func throwFatalDuplicateKeyError(_ sender: Any) {
+        // Triggers: Fatal error: Duplicate keys of type 'Something' were found in a Dictionary.
+        var dict = [HashableViolation(): "value"]
+
+        // Add plenty of items to the dictionary so it uses both == and hash methods, which will cause the crash.
+        for i in 0..<1_000_000 {
+            dict[HashableViolation()] = "value \(i)"
+        }
+    }
+    
     @IBAction func oomCrash(_ sender: UIButton) {
         highlightButton(sender)
         DispatchQueue.main.async {
@@ -99,5 +116,22 @@ class ErrorsViewController: UIViewController {
                 }
             }
         }
+    }
+}
+
+/// When using this class with a dictionary in Swift, it will cause a crash due to the violation of the Hashable contract.
+/// The Swift dict sees multiple keys that are equal but have different hashes, which it can’t resolve safely. When this
+/// happens, the Swift runtime will crash with the error: "Fatal error: Duplicate keys of type 'HashableViolation' were
+/// found in a Dictionary."
+class HashableViolation: Hashable {
+
+    //  always return true, which means every instance of Something is considered equal.
+    static func == (lhs: HashableViolation, rhs: HashableViolation) -> Bool {
+        return true
+    }
+
+    // Always return a different hash value for each instance so we're violating the Hashable contract.
+    func hash(into hasher: inout Hasher) {
+        hasher.combine(ObjectIdentifier(self))
     }
 }