+++ Samples/visionOS-Swift/visionOS-Swift/VisionOSSwiftApp.swift
@@ -9,7 +9,6 @@ struct VisionOSSwiftApp: App {
             options.dsn = "https://6cc9bae94def43cab8444a99e0031c28@o447951.ingest.sentry.io/5428557"
             options.debug = true
             options.tracesSampleRate = 1.0
-            options.profilesSampleRate = 1.0
             options.attachScreenshot = true
             options.attachViewHierarchy = true