+++ Tests/SentryTests/Transaction/SentryTracerTests.swift
@@ -1046,6 +1046,34 @@ class SentryTracerTests: XCTestCase {
         }
     }
 
+    func testMeasurements_ReadWhileWritingOnDifferentThreads() throws {
+        // Arrange
+        let iterations = 100
+        let unit = MeasurementUnitFraction.percent
+
+        let sut = fixture.getSut()
+        let childSpan = sut.startChild(operation: "operation")
+
+        let dispatchQueue = DispatchQueue(label: "testQueue", attributes: .concurrent)
+        let expectation = XCTestExpectation(description: "Set measurements")
+        expectation.expectedFulfillmentCount = iterations
+
+        // Act && Assert
+        for i in 0..<iterations {
+            dispatchQueue.async {
+                sut.setMeasurement(name: "transaction \(i)", value: 12.0, unit: unit)
+                childSpan.setMeasurement(name: "span \(i)", value: 10.0, unit: unit)
+
+                // We only want to ensure we're not crashing here.
+                // We don't care about the actual values as we test these in other tests.
+                XCTAssertGreaterThanOrEqual(sut.measurements.count, 0)
+                expectation.fulfill()
+            }
+        }
+
+        wait(for: [expectation], timeout: 5.0)
+    }
+
     func testFinish_WithUnfinishedChildren() {
         let sut = fixture.getSut(waitForChildren: false)
         let child1 = sut.startChild(operation: fixture.transactionOperation)